requirements.txt
PyOpenGL>=3.1
glfw>=2.5
imgui[glfw]>=2.0.0
